moved around io and utils functions and files
