+++ aeon/dj_pipeline/qc.py
@@ -3,7 +3,7 @@ import pandas as pd
 import numpy as np
 
 from aeon.io import api as aeon_api
-from aeon.util import plotting as aeon_plotting
+from aeon.utils import plotting as aeon_plotting
 
 from . import acquisition
 from . import get_schema_name
